Refactors KV storage
