+++ transform/company.go
@@ -114,7 +114,7 @@ func (c *company) identificadorMatrizFilial(v string) error {
 	return nil
 }
 
-func newCompany(row []string, l *lookups, kv *badgerStorage, privacy bool) (company, error) {
+func newCompany(row []string, l *lookups, kv kvStorage, privacy bool) (company, error) {
 	var c company
 	c.CNPJ = row[0] + row[1] + row[2]
 	c.NomeFantasia = row[4]
@@ -188,7 +188,7 @@ func newCompany(row []string, l *lookups, kv *badgerStorage, privacy bool) (comp
 	}
 	c.DataSituacaoEspecial = dataSituacaoEspecial
 
-	if err := enrichCompany(kv, &c); err != nil {
+	if err := kv.enrichCompany(&c); err != nil {
 		return c, fmt.Errorf("error enriching company %s: %w", cnpj.Mask(c.CNPJ), err)
 	}
 	return c, nil
